+++ Source Packages/io/rerum/crud/TinyOverwrite.java
@@ -19,15 +19,13 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import net.sf.json.JSONObject;
 import io.rerum.tokens.TinyTokenManager;
-import java.util.List;
-import java.util.Map;
 
 
 /**
  *
  * @author bhaberbe
  */
-public class tinyUpdate extends HttpServlet {
+public class TinyOverwrite extends HttpServlet {
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -39,7 +37,6 @@ public class tinyUpdate extends HttpServlet {
      */
         protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException, Exception {
-        request.setCharacterEncoding("UTF-8");
         response.setHeader("Content-Type", "application/json; charset=utf-8");
         response.setCharacterEncoding("UTF-8");
         TinyTokenManager manager = new TinyTokenManager();
@@ -77,7 +74,7 @@ public class tinyUpdate extends HttpServlet {
                 pubTok = manager.generateNewAccessToken();
             }
             //Point to rerum server v1
-            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/update.action");
+            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/overwrite.action");
             HttpURLConnection connection = (HttpURLConnection) postUrl.openConnection();
             connection.setDoOutput(true);
             connection.setDoInput(true);
@@ -89,10 +86,8 @@ public class tinyUpdate extends HttpServlet {
             connection.connect();
             try{
                 DataOutputStream out = new DataOutputStream(connection.getOutputStream());
-                byte[] toWrite = requestJSON.toString().getBytes("UTF-8");
                 //Pass in the user provided JSON for the body of the rerumserver v1 request
-                //out.writeBytes(requestJSON.toString());
-                out.write(toWrite);
+                out.writeBytes(requestJSON.toString());
                 out.flush();
                 out.close(); 
                 codeOverwrite = connection.getResponseCode();
@@ -102,14 +97,8 @@ public class tinyUpdate extends HttpServlet {
                     sb.append(line);
                 }
                 reader.close();
-                for (Map.Entry<String, List<String>> entries : connection.getHeaderFields().entrySet()) {
-                    String values = "";
-                    for (String value : entries.getValue()) {
-                        values += value + ",";
-                    }
-                    response.setHeader(entries.getKey(), values);
-                }
-            } catch (IOException ex) {
+            }
+            catch(IOException ex){
                 //Need to get the response RERUM sent back.
                 BufferedReader error = new BufferedReader(new InputStreamReader(connection.getErrorStream(),"utf-8"));
                 String errorLine = "";
@@ -119,32 +108,10 @@ public class tinyUpdate extends HttpServlet {
                 error.close();
             }
             connection.disconnect();
-            if(manager.getAPISetting().equals("true")){
-                response.addHeader("Access-Control-Allow-Origin", "*"); //To use this as an API, it must contain CORS headers
-            }
             response.setStatus(codeOverwrite);
             response.getWriter().print(sb.toString());
         }
         
-    }
-
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /**
-     * Handles the HTTP <code>GET</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try {
-            processRequest(request, response);
-        } catch (Exception ex) {
-            Logger.getLogger(tinyUpdate.class.getName()).log(Level.SEVERE, null, ex);
-        }
     }
 
     /**
@@ -161,36 +128,7 @@ public class tinyUpdate extends HttpServlet {
         try {
             processRequest(request, response);
         } catch (Exception ex) {
-            Logger.getLogger(tinyUpdate.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-   /**
-     * Handles the HTTP <code>OPTIONS</code> preflight method.
-     * This should be a configurable option.  Turning this on means you
-     * intend for this version of Tiny Things to work like an open API.  
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doOptions(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try {
-            TinyTokenManager manager = new TinyTokenManager();
-            String openAPI = manager.getAPISetting();
-            if(openAPI.equals("true")){
-                //These headers must be present to pass browser preflight for CORS
-                response.addHeader("Access-Control-Allow-Origin", "*");
-                response.addHeader("Access-Control-Allow-Headers", "*");
-                response.addHeader("Access-Control-Allow-Methods", "*");
-            }
-            response.setStatus(200);
-            
-        } catch (Exception ex) {
-            Logger.getLogger(tinyQuery.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(TinyOverwrite.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
@@ -201,7 +139,7 @@ public class tinyUpdate extends HttpServlet {
      */
     @Override
     public String getServletInfo() {
-        return "Short description";
+        return "Overwrites an object in the RERUM store without versioning.";
     }// </editor-fold>
 
 }
